# packages/client/src/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "get": {
        "description": "Returns a paginated list of users",
        "operationId": "UserController_listUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter users by name or email",
            "schema": {
              "example": "john",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Users not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "List all users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "description": "Creates a new user",
        "operationId": "UserController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create new user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Returns a single user by ID",
        "operationId": "UserController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "description": "Updates an existing user",
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Update user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "description": "Deletes a user",
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Delete user",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Brandon API",
    "description": "API docs for Brandon",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3040",
      "description": "Local"
    }
  ],
  "components": {
    "schemas": {
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique identifier for the user",
            "example": 1
          },
          "fullName": {
            "type": "string",
            "description": "The full name of the user",
            "example": "John Doe"
          },
          "phoneNumber": {
            "type": "string",
            "description": "User's phone number",
            "example": "+33612345678"
          },
          "roomNumber": {
            "type": "string",
            "description": "Room number if the user is a guest",
            "example": "101"
          },
          "specialRequests": {
            "type": "string",
            "description": "Any special requests or notes for this user",
            "example": "Requires vegetarian meals"
          }
        },
        "required": [
          "id",
          "fullName",
          "phoneNumber"
        ]
      },
      "PaginatedUsersResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "description": "Total number of users matching the criteria",
            "example": 100
          },
          "next": {
            "type": "object",
            "description": "URL to the next page",
            "example": "/api/users?page=2",
            "nullable": true
          },
          "previous": {
            "type": "object",
            "description": "URL to the previous page",
            "example": "/api/users?page=1",
            "nullable": true
          },
          "results": {
            "description": "List of user results",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            }
          }
        },
        "required": [
          "count",
          "results"
        ]
      },
      "UserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The full name of the user",
            "example": "John Doe",
            "minLength": 2,
            "maxLength": 100
          },
          "phoneNumber": {
            "type": "string",
            "description": "User's phone number",
            "example": "+33612345678",
            "minLength": 10
          },
          "roomNumber": {
            "type": "string",
            "description": "Room number if the user is a guest",
            "example": "101"
          },
          "specialRequests": {
            "type": "string",
            "description": "Any special requests or notes for this user",
            "example": "Requires vegetarian meals"
          }
        },
        "required": [
          "name",
          "phoneNumber"
        ]
      },
...
    }
  }
}